ml_model/requirements.txt

# Pharma Assistant - requirements.txt 


# ----- CORE API -----
fastapi==0.104.1              # Framework API
uvicorn[standard]==0.24.0     # Serveur ASGI

# ----- AI & LLM -----
openai==1.3.0                 # Client OpenAI GPT
chromadb==0.4.18              # Base vectorielle (INCLUT embeddings)

# ----- DATA & PARSING -----
requests==2.31.0              # Requêtes HTTP
beautifulsoup4==4.12.2        # Parsing HTML (DailyMed)
lxml==4.9.3                   # Parser rapide
pandas==2.1.3                 # Manipulation données
numpy==1.24.3                 # Calculs numériques

# ----- CONFIG & TYPES -----
python-dotenv==1.0.0          # Variables d'environnement
pydantic==2.5.0               # Validation de données

# ----- UTILITIES -----
python-multipart==0.0.6       # Upload fichiers
httpx==0.25.1                 # Client HTTP async
